fix(listings): correct internal import paths after M2.5 domain relocation
